+++ core/language/en/gallery_acp.php
@@ -342,7 +342,7 @@ $lang = array_merge($lang, [
 	'RECENT_ON_INDEX'          => 'Show recent images',
 	'RECENT_ON_INDEX_EXP'      => 'Should recent images be displayed on index',
 	'RECENT_ON_INDEX_COUNT'    => 'Recent images count',
-	'FORUM_INDEX_IMAGES'       => 'Gallery images on forum index',
+	'FORUM_INDEX_IMAGES'       => 'Forum index — Gallery image blocks',
 
 	'PHPBB_INTEGRATION' => 'phpBB integration',
 	'PNG_ALLOWED'       => 'Allowed to upload PNG files',